refactor: replace custom analytics URL handling with AnalyticsWithBasePath component
