+++ docs/src/app/layout.tsx
@@ -1,6 +1,6 @@
+import { AnalyticsWithBasePath } from '@/components/analytics';
 import CustomSearchDialog from '@/components/search-dialog';
 import './global.css';
-import { Analytics, type BeforeSendEvent } from '@vercel/analytics/next';
 import { RootProvider } from 'fumadocs-ui/provider/next';
 import { Metadata } from 'next';
 import { Inter } from 'next/font/google';
@@ -17,31 +17,6 @@ export const metadata: Metadata = {
   },
 };
 
-const BASE_PATH = '/docs';
-
-/**
- * Fix analytics URL to include basePath.
- *
- * When the docs site is proxied behind the main site (www.better-stack.ai/docs/*),
- * Next.js internally strips the basePath from URLs. This causes @vercel/analytics
- * to report paths like "/plugins/blog" instead of "/docs/plugins/blog".
- *
- * This beforeSend handler ensures the basePath is included in tracked URLs.
- */
-function fixAnalyticsBasePath(event: BeforeSendEvent): BeforeSendEvent {
-  try {
-    const url = new URL(event.url);
-    // Only add basePath if the pathname doesn't already include it
-    if (!url.pathname.startsWith(BASE_PATH)) {
-      url.pathname = `${BASE_PATH}${url.pathname}`;
-      return { ...event, url: url.toString() };
-    }
-  } catch {
-    // If URL parsing fails, return event unchanged
-  }
-  return event;
-}
-
 export default function Layout({ children }: { children: ReactNode }) {
   return (
     <html lang="en" className={inter.className} suppressHydrationWarning>
@@ -51,7 +26,7 @@ export default function Layout({ children }: { children: ReactNode }) {
             SearchDialog: CustomSearchDialog,
           }}
         >{children}</RootProvider>
-        <Analytics beforeSend={fixAnalyticsBasePath} />
+        <AnalyticsWithBasePath />
       </body>
     </html>
   );
